test(memory): align e2e tests with direct per-turn storage flow
The runtime changed from LLM-based extraction every N turns to direct
per-turn chunk storage (no LLM call). Update tests to match: remove
extraction terminology, drop ExtractionTriggerTest, remove unused
extraction_batch_size config, and simplify store-then-query patterns.
