+++ e2e/testing/memory_tests/__init__.py
@@ -1,0 +1,26 @@
+"""Memory features E2E test package.
+
+Re-exports all test classes so the runner can import them from one place.
+"""
+
+from memory_tests.base import MemoryFeaturesTest, MilvusVerifier
+from memory_tests.test_isolation import UserIsolationTest
+from memory_tests.test_pipeline import (
+    MemoryContentIntegrityTest,
+    MemoryInjectionPipelineTest,
+    SimilarityThresholdTest,
+    StaleMemoryTest,
+)
+from memory_tests.test_storage import MemoryStorageTest, PluginCombinationTest
+
+__all__ = [
+    "MemoryContentIntegrityTest",
+    "MemoryFeaturesTest",
+    "MemoryInjectionPipelineTest",
+    "MemoryStorageTest",
+    "MilvusVerifier",
+    "PluginCombinationTest",
+    "SimilarityThresholdTest",
+    "StaleMemoryTest",
+    "UserIsolationTest",
+]
